% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights.R
\name{parse_insights}
\alias{parse_insights}
\title{Parses the New Relic Insights and convert to an R data structure.}
\usage{
parse_insights(json, include.unknown = FALSE)
}
\arguments{
\item{json}{A character vector containing the JSON to parse.}

\item{include.unknown}{When \code{TRUE} should `unknownGroup` be included in the result.}
}
\value{
data.table generated from \code{json}
}
\description{
Parses the New Relic Insights and convert to an R data structure.
}

