% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights.R
\name{query_insights}
\alias{query_insights}
\title{Queries New Relic Insights using the character vector nrql.}
\usage{
query_insights(nrql, account = Sys.getenv(x = "INSIGHTS_ACCOUNT_ID"),
  key = Sys.getenv(x = "INSIGHTS_ACCOUNT_KEY"), parse = TRUE, ...)
}
\arguments{
\item{nrql}{A character vector containg the NRQL query.}

\item{account}{The New Relic insights account.  Uses the environment variable \code{INSIGHTS_ACCOUNT_ID} as a default.}

\item{key}{The New Relic insights key  Uses the environment variable \code{INSIGHTS_ACCOUNT_KEY} as a default.}

\item{parse}{If \code{FALSE} returns the unparsed JSON}

\item{...}{Passes remaining arguments to \code{\link{parse_insights}}}
}
\description{
Queries New Relic Insights using the character vector nrql.
}

